Replace local MCP with MCPturbo
