+++ genesis_engine/agents/deploy.py
@@ -24,7 +24,8 @@ import shutil
 import platform
 import json
 from pathlib import Path
-from genesis_engine.mcp.agent_base import GenesisAgent, AgentTask, TaskResult
+from genesis_engine.genesis_agent import GenesisAgent
+from genesis_engine.tasking import AgentTask, TaskResult
 
 class DeploymentTarget(str, Enum):
     """Objetivos de despliegue"""
