--- genesis_engine/mcp/__init__.py
@@ -1,3 +1,0 @@
-"""
-Multi-agent Communication Protocol (MCP) para Genesis Engine
-"""
