+++ pyproject.toml
@@ -22,6 +22,8 @@ dependencies = [
     "fastapi>=0.115.14",
     "sqlalchemy>=2.0.23",
     "httpx>=0.24.0",
+    "mcpturbo-core>=2.0.0",
+    "mcpturbo-orchestrator>=1.0.0",
 ]
 
 [project.scripts]
