+++ tests/test_mcp_message_defaults.py
@@ -1,11 +1,8 @@
-from genesis_engine.mcp.message_types import MCPMessage, MessageType
+from mcpturbo_core.messages import Message, MessageType
 
 
 def test_mcp_message_defaults():
-    msg = MCPMessage(sender="src", recipient="dest", action="ping")
+    msg = Message(sender="src")
     assert msg.type == MessageType.REQUEST
-    assert msg.sender_agent == ""
     assert msg.sender == "src"
-    assert msg.recipient == "dest"
-    assert msg.action == "ping"
     assert msg.data == {}
