+++ tests/test_protocol_tasks.py
@@ -1,11 +1,10 @@
 import pytest
-from genesis_engine.mcp.protocol import MCPProtocol
+from mcpturbo_core import MCPProtocol
 
 @pytest.mark.asyncio
 async def test_protocol_tasks_cleanup():
     protocol = MCPProtocol()
     await protocol.start()
+    assert protocol.running
     await protocol.stop()
-    assert protocol.worker_task.done()
-    assert protocol.metrics_task.done()
-    assert protocol.circuit_task.done()
+    assert not protocol.running
